Refactor: move command classes to a subpackage
