+++ src/main/java/com/sparkedhost/accuratereadings/Main.java
@@ -4,6 +4,7 @@ import com.mattmalec.pterodactyl4j.client.entities.Account;
 import com.mattmalec.pterodactyl4j.client.entities.ClientServer;
 import com.mattmalec.pterodactyl4j.exceptions.LoginException;
 import com.mattmalec.pterodactyl4j.exceptions.NotFoundException;
+import com.sparkedhost.accuratereadings.commands.PerformanceCmd;
 import com.sparkedhost.accuratereadings.config.Settings;
 import com.sparkedhost.accuratereadings.managers.PlaceholderAPIManager;
 import com.sparkedhost.accuratereadings.managers.PterodactylManager;
