+++ src/main/java/com/sparkedhost/accuratereadings/commands/PerformanceCmd.java
@@ -1,5 +1,7 @@
-package com.sparkedhost.accuratereadings;
+package com.sparkedhost.accuratereadings.commands;
 
+import com.sparkedhost.accuratereadings.Main;
+import com.sparkedhost.accuratereadings.Methods;
 import com.sparkedhost.accuratereadings.managers.CooldownManager;
 import com.sparkedhost.accuratereadings.managers.PterodactylManager;
 import org.bukkit.Bukkit;
@@ -7,7 +9,6 @@ import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
-import org.bukkit.scheduler.BukkitRunnable;
 
 import java.util.concurrent.TimeUnit;
